requirements.txt
## 6. requirements.txt
requests>=2.31.0
boto3>=1.34.0
python-dotenv>=1.0.0
pathlib2>=2.3.0; python_version < '3.4'